(chore): refactor unittests to one test per file/component (#140)
