+++ src/Components/TimelineNotes/TimelineNotes.test.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import shallowToJSON from 'enzyme-to-json';
 
-import TimelineNotes from '../src/Components/TimelineNotes/TimelineNotes.Component';
-import { testIncidents } from '../mock_endpoints/mockData';
+import TimelineNotes from './TimelineNotes.Component';
+import { testIncidents } from '../../../mock_endpoints/mockData';
 
 describe('Timeline Notes component', () => {
   it('should have all the Timeline Notes content', () => {
