+++ src/actions/ErrorAction.test.js
@@ -1,7 +1,7 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import * as actions from '../src/actions/incidentAction';
-import * as types from '../src/actions/actionTypes';
+import * as actions from './incidentAction';
+import * as types from './actionTypes';
 import moxios from 'moxios';
 
 const middlewares = [thunk];
