+++ src/pages/Dashboard/Dashboard.test.js
@@ -4,14 +4,14 @@ import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import PropTypes from 'prop-types';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import initialState from '../src/reducers/initialState';
+import initialState from '../../reducers/initialState';
 import createRouterContext from 'react-router-test-context';
-import { testIncidents } from '../mock_endpoints/mockData';
-import { Dashboard } from '../src/pages/Dashboard/Dashboard.Component';
-import CircularProgressIndicator from '../src/Components/Progress/Progress.Component';
-import IncidentFilter from '../src/Components/IncidentFilter/IncidentFilter.Component';
-import CustomNotification from '../src/Components/CustomNotification/CustomNotification.Component';
-import CustomMenu from '../src/Components/CustomMenu/CustomMenu.Component';
+import { testIncidents } from '../../../mock_endpoints/mockData';
+import { Dashboard } from './Dashboard.Component';
+import CircularProgressIndicator from '../../Components/Progress/Progress.Component';
+import IncidentFilter from '../../Components/IncidentFilter/IncidentFilter.Component';
+import CustomNotification from '../../Components/CustomNotification/CustomNotification.Component';
+import CustomMenu from '../../Components/CustomMenu/CustomMenu.Component';
 import SelectField from 'material-ui/SelectField';
 
 const middlewares = [thunk];
