+++ src/pages/IncidentTimeline/IncidentTimeline.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { MemoryRouter, Route, Switch } from 'react-router-dom';
 import { mount } from 'enzyme';
 
-import IncidentTimeline from '../src/pages/IncidentTimeline/IncidentTimeline.Component';
+import IncidentTimeline from './IncidentTimeline.Component';
 
 describe('IncidentTimeline component', () => {
   it('renders without crashing', () => {
